Refactor packet_processor module
